requirements.txt
Flask==2.3.3
Flask-WTF==1.1.1
Flask-Login==0.6.3
Werkzeug==2.3.7
WTForms==3.0.1
boto3==1.28.62
python-dotenv==1.0.0
email-validator==2.0.0
Flask-WTF==1.1.1
